data-processing/requirements.txt
# Core
openpyxl>=3.1
pandas>=2.2
PyYAML>=6.0
python-dotenv>=1.0
tqdm>=4.66
yt-dlp>=2025.1.26

# Video / image processing
ffmpeg-python>=0.2.0
numpy>=1.26
opencv-python-headless>=4.10
pillow==11.3.0

# Utilities
matplotlib>=3.9
scikit-learn>=1.5

# Notebooks / app
jupyter>=1.0
notebook>=7.2
streamlit>=1.37

# Retrieval / ML (can be skipped at the very beginning if needed)
# --extra-index-url https://download.pytorch.org/whl/cu121
# torch
# torchvision
# torchaudio
accelerate==1.13.0
einops>=0.8.2
faiss-cpu>=1.8.0
timm>=1.0.0
transformers==4.51.2
ultralytics==8.4.21

# Supabase
supabase>=2.0.0